[RUNTIME] processName sent to stateManager
The name of a destinationProcess is now passed to stateManager along
with any command. This will allow stateManager to choose what process to
send the result to. By default, destinationProcess is the name of the
process putting the command on the stateQueue.

Tests have been changed. BadThingQueue will now print both the
BadThing.event and BadThing.data, separated by a comma.

Refactoring:
new function runtimeSetup() to consolidate all runtime initialization in
a single function.