Initial Commit of Yogibear Firmware
Initial Commit of Yogibear firmware.

Current functionality DOES NOT IMPLIMENT HIBIKE.  It merely matches the
functionality of the arduino prototyping (AkA type magic numbers into
the console & get the motor to do things)

The guts of the firmware were developed on the stanley_master branch of
the yogi-bear-arduino repo, in the "integrated" folder.

That arduino code was then copy-pastaed into pie central, and then
modified so that they would compile.